Implement the complete application
- Completed the api development
- Implemented the portfolio web site
- Implemented admin dashboard to update the portfolio

+++ models/engines/db_storage.py
@@ -25,21 +25,25 @@ class DBStorage(SQLAlchemy):
             obj = cls(**kwargs)
             self.session.add(obj)
             self.session.commit()
+
+            return self.session.get(cls, obj.id)
         except IntegrityError as err:
             self.session.rollback()
             raise ValueError(f"Database error: {err}")
 
-        return obj
-
-    def remove_object(self, obj: ModelType) -> None:
+    def remove_object(self, cls: t.Type[ModelType], id: str) -> None:
         try:
+            obj = self.session.get(cls, id)
+            if not obj:
+                raise ValueError('id does not exist')
+
             self.session.delete(obj)
             self.session.commit()
         except IntegrityError as err:
             self.session.rollback()
             raise ValueError(f"Database error: {err}")
 
-    def fetch_an_object_by(
+    def fetch_object(
         self,
         cls: t.Type[ModelType],
         **kwargs: t.Mapping
@@ -49,7 +53,7 @@ class DBStorage(SQLAlchemy):
         """
         return self.session.query(cls).filter_by(**kwargs).first()
 
-    def fetch_object_by(
+    def fetch_objects(
         self,
         cls: t.Type[ModelType],
         **kwargs: t.Mapping
@@ -58,6 +62,12 @@ class DBStorage(SQLAlchemy):
         Fetch object(s) from the database using attributes
         """
         return self.session.query(cls).filter_by(**kwargs).all()
+
+    def fetch_all(self, cls: t.Type[ModelType]) -> t.List[ModelType]:
+        """
+        Fetch all the objects of type cls from the database
+        """
+        return self.session.query(cls).all()
 
     def update_object(
         self,
@@ -68,14 +78,25 @@ class DBStorage(SQLAlchemy):
         """
         Update a given model object
         """
+        from datetime import datetime
+        from hashlib import md5
+
         obj = self.session.get(cls, id)
+        if not obj:
+            raise ValueError('id does not exist')
+
         for key, value in kwargs.items():
             if key not in obj.__table__.columns.keys():
                 raise ValueError(
                     f"{key} field does not exist in {cls.__name__} Model"
                 )
 
+            if key == 'password':
+                password_hash = md5(value.encode('utf-8'))
+                value = password_hash.hexdigest()
+
             setattr(obj, key, value)
+            obj.updated_at = datetime.now()
 
         try:
             self.session.add(obj)
